Duration and DurationTime are now separate files
Duration and DurationTime are now moved into theit own files.
Therefore the tests also had to be split into two different files.
The logic is still the same.

time base snapshot filters - minor corrections

internal/data/duration_time.go: stricter checking for logic errors

internal/data/snapshot_find.go: rewording of some comments
changed to 'if err := ...; err != nil' clause

internal/data/snapshot_find_test.go:
removed duplicate tests for data.DurationTime
